requirements.txt
pip==22.0.4
django==3.2
python-decouple==3.6
psycopg2-binary==2.9.3
celery==5.2.3
flower==1.0.0
tweepy==4.5.0
django-redis==5.2.0
requests==2.27.1
pillow==9.0.1
numpy==1.22.2
arabic-reshaper==2.1.3
python-bidi==0.4.2
gunicorn==20.1.0
